プレイヤークラスをインゲームに追加

インゲームのオンコリジョン関数内に棘に関する当たり判定をregionでまとめた

棘とプレイヤーとの当たり判定の処理の追加

棘と棘同士の当たり判定の処理を追加

棘と壁との判定を仮で追加

+++ Game/map.cpp
@@ -56,7 +56,7 @@ void Map::Update() {
 
 
 
-bool Map::IsHitWall(const Vector3& playerpos,const int& wide) {
+bool Map::IsHitWall(const Vector3& playerpos,const float& wide) {
 	if ((playerpos.x+wide)>wallX.max) {
 		return true;
 	}
